Correct errors and optimize details.
