--- week2/static/css/attraction.css
@@ -1,379 +1,0 @@
-body {
-	margin: 0;
-	padding: 0;
-  font-family: 'Noto Sans TC';
-  font-size: 16px;
-}
-
-main{
-  margin: auto;
-  width: 1200px;
-  height: auto;
-  margin-bottom: 40px;
-}
-
-footer {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	background-color: #757575;
-  width: 100%;
-	height: 100px;
-  color: #FFFF;
-  position: static;
-}
-
-button{
-  border: 0;
-}
-
-.header__hr{
-  border: 1px solid #E8E8E8;
-  width: 100%; 
-  margin: 0;
-  margin-bottom: 40px;
-}
-
-.main__hr{
-  margin-top: 20px;
-  border: 1px solid #E8E8E8;
-  width: 100%; 
-  margin: 0;
-}
-
-.header {
-  display: flex;
-  width: 1200px;
-  margin: auto;
-  justify-content: space-between;
-  align-items: center;
-  height: 54px;
-}
-
-.header__title {
-  display: flex;
-  font-size: 30px;
-  color: #448899;
-  font-weight: 700;
-  align-items: center;
-  height: 34px;
-}
-
-.header__nav {
-  display: flex;
-  gap:20px
-}
-
-.header__button {
-  all: unset;
-  cursor: pointer;
-  color: #666666;
-}
-
-.main__grid{
-  display: grid;
-  grid-template-columns: repeat(2, 1fr );
-  width: 100%;
-  height: 100%;
-  gap: 30px;
-}
-
-.main__grid> *:nth-child(3){
-  grid-column: span 2; 
-}
-
-.main__grid> *:nth-child(4){
-  grid-column: span 2; 
-}
-
-.attraction__element__leftArrow{
-  width: 36px;
-  height: 36px;
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%); 
-  left: 10px;
-  cursor:pointer;
-}
-
-.attraction__element__rightArrow{
-  width: 36px;
-  height: 36px;
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%); 
-  right: 10px;
-  cursor:pointer;
-}
-
-.attraction__element__image {
-  position: relative;
-  border-radius: 5px;
-  display: flex;
-  width: 100%;
-  height: 100%;
-}
-
-.attraction__element__slide {
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-  border-radius: 5px;
-  height:100%;
-  width:100%;
-}
-
-.attraction__element__dots {
-  display: flex;
-  position: absolute;
-  bottom: 10px;
-  left: 50%;
-  transform: translateX(-50%);
-  max-width: 80%; 
-  gap: 10px; 
-}
-
-.attraction__element__dot {
-  border-radius: 50%;
-  height: 8px; 
-  width: 8px; 
-  background-color: rgba(255, 255, 255, 0.7); 
-  cursor: pointer;
-  flex-shrink: 0;; 
-}
-
-.attraction__element__blackDot{
-  background-color: black;
-  border-radius: 50%;
-  border: 1px solid white;
-}
-
-.attraction__element__order{
-  height: 100%;
-  width: 100%;
-  display: grid;
-  grid-template-rows: 33px 22px 295px;
-  gap: 20px;
-  margin: auto;
-}
-
-.order__element__name{
-  height: 33px;
-  font-size: 24px;
-  font-weight: 700;
-  color: #666666;
-  margin-bottom: 0;
-  align-items: center;
-}
-
-.order__element__category{
-  height: 22px;
-  font-size: 16px;
-  font-weight: 400;
-  color: #666666;
-  align-items: center;
-}
-
-.order__element{
-  background-color: #E8E8E8;
-  display: grid;
-  padding: 20px;
-  padding-top: 15px;
-  border-radius: 5px;
-  align-items: center;
-}
-
-.order__element > div:nth-child(2){
-  color: #666666;
-  font-size: 16px;
-  font-weight: 400;
-  align-items: center;
-}
-
-.order__element > div{
-  color: #666666;
-  font-size: 16px;
-  font-weight: 700;
-  align-items: center;
-}
-
-.order__element > button{
-  color: #FFFF;
-  font-size: 19px;
-  font-weight: 400;
-  background-color:#448899;
-  border-radius: 5px;
-  width: 150px;
-  height: 36px;
-  align-items: center;
-}
-
-.describe__element{
-  display: grid;
-  margin: auto;
-  width: 1200px;
-  gap:30px;
-
-  margin-bottom: 40px;
-}
-
-.describe__element__introduce{
-  color:#666666;
-  font-size: 16px;
-  font-weight: 400;
-  height: auto;
-  text-align: justify;
-  margin: 0;
-}
-
-.describe__element__address{
-  color:#666666;
-  font-size: 16px;
-  font-weight: 400;
-  text-align: justify;
-}
-
-.describe__element__transportation{
-  color:#666666;
-  font-size: 16px;
-  font-weight: 400;
-  text-align: justify;
-}
-
-
-input[type="date"] {
-  position: relative;
-  background: url("/static/imgs/image.png") no-repeat right 10px center;
-  background-size: 20px;
-  padding-right: 30px;
-  border: 0;;
-  height: 30px;
-  background-color: #FFFFFF;
-}
-
-input[type="date"]::-webkit-calendar-picker-indicator {
-  opacity: 0;
-  position: absolute;
-  right: 10px;
-  width: 100%;
-  height: 100%;
-  cursor: pointer;
-}
-
-input[type="radio"] {
-  appearance: none;
-  width: 18px;
-  height: 18px;
-  background-color: #FFFFFF;
-  border-radius: 50%;
-  position: relative;
-}
-
-input[type="radio"]:checked {
-  background-color: #448899;
-  border: 2px solid white;
-}
-
-.err__element{
-  margin: auto;
-  width: 1200px;
-  font-size: 40px;
-} 
-
-
-
-
-@media screen and (min-width: 601px) and (max-width: 1201px) {
-  .header{
-    width: 90%;
-  }
-  
-  main{
-    width: 100%;
-  }
-
-  .main__grid{
-    grid-template-columns: repeat(1, 1fr); 
-  }
-  
-  .main__grid > *:nth-child(1){
-    height: 360px;
-    width: 100%;
-    grid-column: span 2; 
-  }
-  
-  .main__grid > *:nth-child(2),
-  .main__grid > *:nth-child(4){
-    width: 90%;
-    grid-column: span 2; 
-  }
-  
-  .attraction__element__slide{
-    border-radius: 0;
-  }
-  .describe__element{
-    width: 100%;
-  }
-
-  .main__hr{
-    width: 90%; 
-    margin: auto;
-  }
-  
-  .attraction__element__order{
-    grid-template-rows: auto;
-  }
-
-  .order__element{
-    gap: 20px;
-  }
-
-  footer{
-    position: static;
-  }
-}
-
-@media screen and (min-width: 360px) and (max-width:600px) {
-  .header{
-    width: 90%;
-  }
-  
-  main{
-    width: 100vw;
-  }
-
-  .main__grid{
-    grid-template-columns: repeat(1, 1fr); 
-  }
-  
-  .main__grid > *:nth-child(1){
-    height: 360px;
-    width: 100%;
-    grid-column: span 2; 
-  }
-  
-  .main__grid > *:nth-child(2),
-  .main__grid > *:nth-child(4){
-    width: 90%;
-    grid-column: span 2; 
-  }
-
-  .attraction__element__slide{
-    border-radius:0px;
-  }
-
-  footer{
-    position: static;
-  }
-
-  .header__hr{
-    margin-bottom: 0px;
-  }
-  
-  .main__hr{
-    width: 90%; 
-    margin: auto;
-  }
-  
-}
-
