--- week2/static/css/index.css
@@ -1,272 +1,0 @@
-body {
-	margin: 0;
-	padding: 0;
-  font-family: 'Noto Sans TC';
-  font-size: 16px;
-}
-
-main {
-	flex-grow:1;
-}
-
-footer {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	background-color: #757575;
-	height: 100px;
-  color: #FFFF;
-}
-
-.header {
-  display: flex;
-  width: 1200px;
-  margin: auto;
-  justify-content: space-between;
-  align-items: center;
-  height: 54px;
-}
-
-.header__title {
-  display: flex;
-  font-size: 30px;
-  color: #448899;
-  font-weight: 700;
-  align-items: center;
-  height: 34px;
-}
-
-.header__nav {
-  display: flex;
-  gap:20px
-}
-
-.header__button {
-  all: unset;
-  cursor: pointer;
-  color: #666666;
-}
-
-.search__section {
-  background-image: url(/static/imgs/hero_section.png);
-  background-size: cover;
-  background-position: center;
-  height: 320px;
-  width: 100%;
-  display: flex;
-  margin-bottom: 40px;
-}
-
-.search__section__content {
-  width: 1200px;
-  height: 169px;
-  margin: auto;
-}
-
-.search__heading__H2{
-  color: #f8f8f8;
-	font-size: 28px;
-  letter-spacing: 2px;
-	width: 100%;
-}
-
-.search__heading__H4 {
-  color: #f8f8f8;
-  letter-spacing: 1px;
-	width: 100%;
-}
-
-.search {
-  display: flex;
-  align-items: center;
-	width: 100%;
-}
-
-.search__input {
-  display: flex;
-  padding-top:5px ;
-  color: #757575;
-  font-weight: 700;
-  width: 400px;
-  height: 44px;
-  border-radius: 5px 0 0 5px;
-  border: none;
-}
-
-.search__button {
-  background-color: #448899;
-  height: 46px;
-  width: 60px;
-  border: none;
-  border-radius: 0 5px 5px 0;
-}
-
-.search__button > img {
-  height: 30px;
-  width: 30px;
-  cursor: pointer;
-}
-
-.mrt__title__container {
-  display: flex;
-  width: 1200px;
-  margin: auto auto 40px auto;
-}
-
-.mrt__title {
-  display: flex;
-  gap: 10px;
-	width: 100%;
-  flex-wrap: nowrap;
-  overflow-x: auto;
-  white-space: nowrap;
-  scroll-behavior: smooth;
-}
-
-.mrt__title::-webkit-scrollbar {
-  display: none;
-}
-
-.mrt__button {
-  background-color: #FFFF;
-  color: #666666;
-  border: none;
-  cursor: pointer;
-  font-weight: 700;
-  font-size:16px;
-}
-
-.mrt__button:hover{
-  color: #000000;
-}
-
-.arrow--left {
-  left: 0;
-  padding-right: 20px;
-  cursor: pointer;
-}
-
-.arrow--right {
-  right: 0;
-  padding-left: 20px;
-  cursor: pointer;
-}
-
-.grid {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: 30px;
-  width: 1200px;
-  margin: auto;
-	margin-bottom: 30px;
-}
-
-.attractions__element {
-  position: relative;
-  border: 1px solid #E8E8E8;
-	height: 200px;
-  border-radius: 5px;
-}
-
-.attractions__element__image {
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-  height: 80%;
-  width: 100%;
-  border-radius: 5px 5px 0 0;
-}
-
-.attractions__element__name {
-  background-color: rgba(0, 0, 0, 0.6);
-  color:#FFFF;
-  height: 20px;
-	font-weight: 700;
-  padding: 10px;
-  width: calc(100% - 20px);
-  position: absolute;
-  bottom: 20%;
-	white-space: nowrap;      
-  overflow: hidden;         
-  text-overflow: ellipsis; 
-}
-
-.attractions__element__bottom {
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  height: 40px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.attractions__element__bottom__left {
-  padding-left: 10px;
-  font-weight: 700;
-	color: #757575;
-}
-
-.attractions__element__bottom__right {
-  padding-right: 10px;
-  font-weight: 700;
-	color: #757575;
-}
-
-
-@media screen and (min-width: 601px) and (max-width: 1201px) {
-	.header{
-		width: 90%;
-	}
-
-
-	.search__section__content {
-		width: 90%;
-		height: 169px;
-		margin: auto;
-	}
-
-	.grid {
-		width: 90%;
-		grid-template-columns: repeat(2, 1fr);
-	}
-
-  .mrt__title{
-    -webkit-overflow-scrolling: touch;
-  }
-  
-	.mrt__title__container {
-		width: 90%;
-	}
-
-}
-
-
-@media screen and (min-width: 360px) and (max-width:600px) {
-	.header{
-		width: 90%;
-	}
-
-
-	.search__section__content {
-		width: 90%;
-		height: 169px;
-		margin: auto;
-	}
-
-	.grid {
-    grid-template-columns: repeat(1, 1fr); 
-		width: 90%;
-  }
-
-	.mrt__title__container {
-		width: 90%;
-	}
-
-
-	.search__section {
-		height: 360px;
-		background-image: url(/static/imgs/phone.png);
-	}
- 
-}
