--- week2/static/index.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" type="text/css" href="/static/css/index.css">
-  <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+TC:wght@100..900&display=swap" rel="stylesheet">
-  <title>Index</title>
-  
-</head>
-<body>
-  <header class="header">
-    <p class="header__title">台北一日遊</p>
-    <nav class="header__nav">
-      <button class="header__button">預定行程</button>
-      <button class="header__button">登入/註冊</button>
-    </nav>
-  </header>
-  <main>
-    <div class="search__section">
-      <div class="search__section__content">
-        <h2 class="search__heading__H2">輕鬆享受台北一日悠閒</h2>
-        <h4 class="search__heading__H4">探索每個角落，體驗城市的深度旅遊行程</h4>
-        <div class="search">
-          <input type="text" class="search__input" placeholder="   輸入景點名稱查詢">
-          <button type="submit" class="search__button">
-            <img src="/static/imgs/icon_search.png" alt="搜尋" />
-          </button>
-        </div>
-      </div>
-    </div>
-
-    <div class="mrt__title__container">
-      <img src="/static/imgs/arrow left.png" class="arrow--left" alt="Left Arrow">
-      <div class="mrt__title"></div>
-      <img src="/static/imgs/arrow right.png" class="arrow--right" alt="Right Arrow">
-    </div>
-    <div class="grid"></div>
-  </main>
-  <footer>
-    COPYRIGHT © 2021 台北一日遊
-  </footer>
-  <script src="/static/js/index.js"></script>
-</body>
-</html>
